+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.0.1] - 2023-07-17
+### Fixed
+- Fixed mobile phone number get from Order address
+- Fixed delivery method recognition
+- Changed tracking_number length in database to 32 chars
+
+### Improved
+- Added Readme file
+
 ## [1.0.0] - 2023-01-13
 ### Init
 - Initial launch of the module